requirements.txt
python-telegram-bot==22.6

aiosqlite>=0.20.0

python-dotenv>=1.0.0

httpx>=0.23.0,<1.0.0
